adds highlight to menu bar options and modal lists, makes it so that current app text changes per app, in dock now light up when apps open
