+++ script.js
@@ -1,98 +1,16 @@
-const updateTime = () => {
-	setInterval("time()", 1000);
-};
+import {
+	addPhotosClasses,
+	addTexteditClasses,
+	addRemindersClasses,
+} from "./modules/add-classes.js";
+import {
+	removePhotosClasses,
+	removeTexteditClasses,
+	removeRemindersClasses,
+} from "./modules/remove-classes.js";
+import { time } from "./modules/clock-functions.js";
 
-const time = () => {
-	let date = new Date();
-	let day = () => {
-		switch (date.getDay()) {
-			case 0:
-				return "Sun";
-				break;
-			case 1:
-				return "Mon";
-				break;
-			case 2:
-				return "Tues";
-				break;
-			case 3:
-				return "Wed";
-				break;
-			case 4:
-				return "Thurs";
-				break;
-			case 5:
-				return "Fri";
-				break;
-			case 6:
-				return "Sat";
-				break;
-		}
-	};
-	let month = () => {
-		switch (date.getMonth()) {
-			case 0:
-				return "Jan";
-				break;
-			case 1:
-				return "Feb";
-				break;
-			case 2:
-				return "Mar";
-				break;
-			case 3:
-				return "Apr";
-				break;
-			case 4:
-				return "May";
-				break;
-			case 5:
-				return "Jun";
-				break;
-			case 6:
-				return "Jul";
-				break;
-			case 7:
-				return "Aug";
-				break;
-			case 8:
-				return "Sep";
-				break;
-			case 9:
-				return "Oct";
-				break;
-			case 10:
-				return "Nov";
-				break;
-			case 11:
-				return "Dec";
-				break;
-		}
-	};
-	const minutes = () => {
-		const minute = date.getMinutes();
-		if (parseInt(minute) < 10) {
-			return "0" + minute;
-		} else {
-			return minute;
-		}
-	};
-
-	const hours = () => {
-		const hour = date.getHours();
-		if (parseInt(hour) < 10) {
-			return "0" + hour;
-		} else {
-			return hour;
-		}
-	};
-
-	let dayOfMonth = date.getUTCDate();
-
-	let currentDateTime = `${day()} ${dayOfMonth} ${month()} ${hours()}:${minutes()}`;
-	let clockSpan = document.getElementById("clock");
-	clockSpan.innerHTML = currentDateTime;
-};
+setInterval(time, 1000);
 
 const menubarModals = document.querySelectorAll(".menu-bar__modal");
 
@@ -110,6 +28,7 @@ const overlay = document.querySelector(".overlay");
 
 menubarButtons.forEach((btn) => {
 	btn.addEventListener("click", () => {
+		btn.classList.add("menu-bar__btns__item--active");
 		overlay.classList.add("overlay__active");
 		if (btn.classList.contains("menu-bar__btns__item--apple-logo")) {
 			apple.classList.add("menu-bar__modal__apple__active");
@@ -133,6 +52,9 @@ menubarButtons.forEach((btn) => {
 
 overlay.addEventListener("click", () => {
 	overlay.classList.remove("overlay__active");
+	menubarButtons.forEach((btn) => {
+		btn.classList.remove("menu-bar__btns__item--active");
+	});
 	menubarModals.forEach((modal) => {
 		if (modal.classList.contains("menu-bar__modal__apple__active")) {
 			modal.classList.remove("menu-bar__modal__apple__active");
@@ -156,71 +78,28 @@ overlay.addEventListener("click", () => {
 	});
 });
 
-const appModals = document.querySelectorAll(".app-modal");
 const desktopIcons = document.querySelectorAll(".desktop__icon");
-
-console.log(appModals);
-console.log(desktopIcons);
+const currentAppText = document.querySelector(
+	".menu-bar__btns__item--current-app"
+);
 
 desktopIcons.forEach((icon) => {
 	icon.addEventListener("dblclick", () => {
 		if (icon.classList.contains("photos")) {
-			document
-				.querySelector(".app-modal__photos")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.add("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.add("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.remove("app-modal__top-bar__textedit--active");
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.remove("app-modal__active");
+			currentAppText.innerText = "Photos";
+			addPhotosClasses();
+			removeTexteditClasses();
+			removeRemindersClasses();
 		} else if (icon.classList.contains("textedit")) {
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.add("app-modal__top-bar__textedit--active");
-			document
-				.querySelector(".app-modal__photos")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.remove("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.remove("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.remove("app-modal__active");
+			currentAppText.innerText = "TextEdit";
+			addTexteditClasses();
+			removePhotosClasses();
+			removeRemindersClasses();
 		} else if (icon.classList.contains("reminders")) {
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__photos")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.remove("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.remove("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.remove("app-modal__top-bar__textedit--active");
+			currentAppText.innerText = "Reminders";
+			addRemindersClasses();
+			removePhotosClasses();
+			removeTexteditClasses();
 		}
 	});
 });
@@ -230,62 +109,20 @@ const dockIcons = document.querySelectorAll(".dock__app");
 dockIcons.forEach((icon) => {
 	icon.addEventListener("click", () => {
 		if (icon.classList.contains("photos")) {
-			document
-				.querySelector(".app-modal__photos")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.add("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.add("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.remove("app-modal__top-bar__textedit--active");
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.remove("app-modal__active");
+			currentAppText.innerText = "Photos";
+			addPhotosClasses();
+			removeTexteditClasses();
+			removeRemindersClasses();
 		} else if (icon.classList.contains("textedit")) {
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.add("app-modal__top-bar__textedit--active");
-			document
-				.querySelector(".app-modal__photos")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.remove("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.remove("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.remove("app-modal__active");
+			currentAppText.innerText = "TextEdit";
+			addTexteditClasses();
+			removePhotosClasses();
+			removeRemindersClasses();
 		} else if (icon.classList.contains("reminders")) {
-			document
-				.querySelector(".app-modal__reminders")
-				.classList.add("app-modal__active");
-			document
-				.querySelector(".app-modal__photos")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.remove("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.remove("app-modal--photos--active");
-			document
-				.querySelector(".app-modal__textedit")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.remove("app-modal__top-bar__textedit--active");
+			currentAppText.innerText = "Reminders";
+			addRemindersClasses();
+			removePhotosClasses();
+			removeTexteditClasses();
 		}
 	});
 });
@@ -293,43 +130,38 @@ dockIcons.forEach((icon) => {
 const closeBtns = document.querySelectorAll(
 	".app-modal__top-bar__btns__btn--close"
 );
-console.log(closeBtns);
 
 closeBtns.forEach((btn) => {
 	btn.addEventListener("click", () => {
 		if (btn.classList.contains("photos1")) {
+			currentAppText.innerText = "Finder";
+			removePhotosClasses();
 			document
-				.querySelector(".app-modal__photos")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar--photos")
-				.classList.remove("app-modal__top-bar--photos--active");
-			document
-				.querySelector(".app-modal--photos")
-				.classList.remove("app-modal--photos--active");
+				.querySelector("#dotPhotos")
+				.classList.remove("dock__app__dot--app-open");
 		} else if (btn.classList.contains("textedit")) {
+			currentAppText.innerText = "Finder";
+			removeTexteditClasses();
 			document
-				.querySelector(".app-modal__textedit")
-				.classList.remove("app-modal__active");
-			document
-				.querySelector(".app-modal__top-bar__textedit")
-				.classList.remove("app-modal__top-bar__textedit--active");
+				.querySelector("#dotTextedit")
+				.classList.remove("dock__app__dot--app-open");
 		} else if (btn.classList.contains("reminders")) {
+			currentAppText.innerText = "Finder";
+			removeRemindersClasses();
 			document
-				.querySelector(".app-modal__reminders")
-				.classList.remove("app-modal__active");
+				.querySelector("#dotReminders")
+				.classList.remove("dock__app__dot--app-open");
 		}
 	});
 });
 
 const reminderForm = document.querySelector("#reminderForm");
-const reminderInput = document.querySelector("#reminderInput");
 
 const reminderList = document.querySelector(
 	".app-modal--reminders--reminder-section__list-container__reminder-list"
 );
 
-console.log(reminderList.length);
+let remindersDone = document.querySelectorAll("#remindersDone");
 
 reminderForm.addEventListener("submit", (event) => {
 	event.preventDefault();
@@ -338,21 +170,23 @@ reminderForm.addEventListener("submit", (event) => {
 	if (reminder) {
 		const newListItem = reminderList.appendChild(document.createElement("li"));
 		newListItem.innerHTML =
-			'<button class="app-modal--reminders--reminder-section__list-container__reminder-list__item__done-button"></button>' +
+			'<button class="app-modal--reminders--reminder-section__list-container__reminder-list__item__done-button" id="remindersDone"></button>' +
 			reminder;
 		newListItem.classList.add(
 			"app-modal--reminders--reminder-section__list-container__reminder-list__item"
 		);
 	}
 	document.querySelector("#reminderInput").value = "";
+	remindersDone = document.querySelectorAll("#remindersDone");
+	console.log(remindersDone);
+	btnIndex++;
 });
 
-const remindersDone = document.querySelectorAll("remindersDone");
-
-remindersDone.forEach((btn) => {
+document.querySelectorAll("#remindersDone").forEach((btn) => {
 	btn.addEventListener("click", (event) => {
 		event.preventDefault();
-		const index = reminderList.getElementsByTagName("li").length;
-		console.log(index);
+		event.target.parentNode.remove();
+		remindersDone = document.querySelectorAll("#remindersDone");
+		console.log("btn ppress", remindersDone);
 	});
 });
